FL - 2. Scrape_Website_using_LLM/requirements.txt
scrapegraphai
playwright 
python-dotenv